+++ test/better_test.js
@@ -1,7 +1,11 @@
 var Sql = require("sqlate").Sql
 var Sqlite3 = require("better-sqlite3")
 var SqliteHeaven = require("../better")
-var db = new Sqlite3(":memory:", {memory: true})
+var BETTER_SQLITE3_VERSION = require("better-sqlite3/package").version
+
+var db = /^6\./.test(BETTER_SQLITE3_VERSION)
+	? new Sqlite3(":memory:", {memory: true})
+	: new Sqlite3(":memory:")
 
 function execute(sql) {
 	if (!(sql instanceof Sql)) throw new TypeError("Not Sql: " + sql)